SearchActivity refactor to move inner classes to separate files
