+++ app/src/main/res/layout/activity_search.xml
@@ -49,19 +49,20 @@
             android:src="@drawable/at_the_beach"
             android:contentDescription="@string/scenic_photo"
             android:tint="@color/homeBackgroundTint" />
-        <TextView
+        <!--TextView
             android:id="@android:id/empty"
             android:layout_height="@dimen/no_results_height"
             android:layout_width="match_parent"
             android:textAlignment="center"
             android:textSize="@dimen/no_results_text_size"
             android:text="@string/no_results_msg"
-            android:textColor="@color/cardTextColor"/>
+            android:textColor="@color/cardTextColor"/-->
         <androidx.recyclerview.widget.RecyclerView android:id="@+id/recyclerView"
             android:layout_height="match_parent"
             android:layout_width="match_parent"
             android:divider="@color/listDividerColor"
             android:dividerHeight="@dimen/note_list_divider_height"
-            android:scrollbarThumbVertical="@drawable/scrollbar_vertical_thumb"/>
+            android:scrollbarThumbVertical="@drawable/scrollbar_vertical_thumb"
+            app:layoutManager="LinearLayoutManager"/>
     </RelativeLayout>
 </LinearLayout>
